refactor: move stats into algorithm crate & response into api crate
