+++ server/algorithms/src/lib.rs
@@ -4,4 +4,5 @@ pub mod bootstrapping;
 pub mod clustering;
 pub mod routing;
 pub mod s2;
+pub mod stats;
 pub mod utils;
