+++ server/api/src/private/geofence_project.rs
@@ -3,7 +3,10 @@ use super::*;
 use serde::Deserialize;
 use serde_json::json;
 
-use crate::model::{api::args::Response, db::geofence_project, KojiDb};
+use crate::{
+    model::{db::geofence_project, KojiDb},
+    utils::response::Response,
+};
 
 #[get("/all/")]
 async fn get_all(conn: web::Data<KojiDb>) -> Result<HttpResponse, Error> {
