+++ server/model/src/api/mod.rs
@@ -18,7 +18,6 @@ pub mod point_struct;
 pub mod poracle;
 pub mod single_struct;
 pub mod single_vec;
-pub mod stats;
 pub mod text;
 
 pub type Precision = f64;
